process capability; resolution time; QFD
The Six Sigma is an innovative, systematic, data-driven methodology for improving processes that reduce waste and defects/errors. The dynamic measure within Six Sigma is the number of defects and identifying the causes for variation. The variations are agile that creates uncertainty in delivering the desired outcome. Service organizations are offering services at a predetermined level stated in SLA. This paper attempts to standardize, reduce variation in offering service quality through identifying the causes of variations in the project operations by establishing specifications. The study shows the deviations among the projects teams. Analysing the time-variation in resolving the problems encountered in different projects, variation in the projects lapsed the delivery content within the SLA. Results after introducing process capabilities and online auto-triggering system enabled, variation in resolution times dropped. Team members were redeployed to other projects thus time-variations is controlled with multiple business opportunities.