Using the Web to Develop Global Ethical Engineering Students
Ethical engineering practice is a global issue. However, cultural norms and social realities may result in differences in ethical behavior. A basic instructional challenge is developing the ability of students to understand ethical practice and to facilitate discussion of ethical issues across regional and cultural boundaries. The present project seeks to facilitate discussion and analysis of ethical practices of undergraduate engineering students at our university, Texas Tech, and peers at universities in Ukraine and India, our partnering countries. An undergraduate ethics course for engineering majors and the website https://EthicalEngineer. ttu. edu are currently the primary channels through which this project is being developed and implemented. The paper describes i) steps in developing the website, which is currently operational, ii) progress developing partnerships and recruiting participants, and iii) results from preliminary tests of machine-assisted methods for analyzing students’ submissions to the website. The Ethical Engineer is an emerging web-based innovation for facilitating cross-cultural discourse in engineering ethics.