Feasibility of roof top rainwater harvesting potential - A case study of South Indian University
Shortage of water source has been a major problem for rapidly growing cities in country like India due to increase in consumption of water. Besides this the coastal cities like Mangaluru are currently challenged by temporary floods due to increased precipitation. Hence it demands a flawless planning for managing the water resource. Rooftop Rainwater Harvesting (RTRWH) has proven to be the most economical and environmental friendly method. The principal objective of this study is to design the RWH system for a South Indian University (SIU), to transform the present campus with green initiative by effective utilization of water resource. An integrated study on rooftop storm water runoff for quantitative and qualitative analysis was conducted. The SIU can resolve the water scarcity issue by accumulating about 1,13,678.9 m3 of stormwater from rooftop in a year, and use it during the non-monsoon season. These findings can significantly influence the similar RTRWH initiatives in various government and private establishments for sustainable water management.